% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PS_GDD.R
\name{PS_GDD}
\alias{PS_GDD}
\title{Growing Degree Days for Philaenus spumarius eggs}
\usage{
PS_GDD(T)
}
\arguments{
\item{T}{Temperature in Celsius. No default.}
}
\description{
This function allows to compute the Growing Degree Days for the latter computation of Philaenus spumarius hatching probabilities.
}
\examples{
PS_GDD(14.5)
}
\keyword{GDD,}
\keyword{Temperature}
