% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PS_hatching_probability.R
\name{PS_hatching_probability}
\alias{PS_hatching_probability}
\title{Hatching probability for Philaenus spumarius eggs based on GDDs}
\usage{
PS_hatching_probability(GDD)
}
\arguments{
\item{GDD}{Growing Degree Days for Philaenus spumarius. No default.}
}
\description{
This function allows to compute the Philaenus spumarius eggs hatching probabilities.
}
\examples{
PS_hatching_probability(280)
}
\keyword{GDD,}
\keyword{Probabilities}
